requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pandas
openai
chromadb
python-dotenv
sentence-transformers